Add hooks to override types when parsing verilog
Added a general "type_map" argument which allows users to override
parsed types. This commit just does the plumbing to propagate this
argument from the top level functions.

The motivation for this change is to allow us to parse specific ports
as clocks, since in verilog everything is a "Bit".