WIP - Initial logic for m.circuit.sequential
Summary: This is an initial pass at the logic for m.circuit.sequential.
The intended design is to parse the `__init__` method to determine the
register declarations from the user. The `__call__` method will be
translated into a `m.circuit.combinational` definition that consumes the
inputs to the circuits and the declared registers (contained in self)
and returns the outputs of the circuits. The compiler logic will
generate the surrounding magma circuit definition including an IO list,
a set of register instantiations, a m.circuit.combinational definition
and invocation, and the wiring of the outputs.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Blame Revision:

Differential Revision: https://phabricator.intern.facebook.com/D13645206