Fixes issues in the new uniquification logic
Major change is to add new `__hash__` and `__eq__` methods for circuits.
Given the new uniquification naming scheme, we can assume that names are
unique and use them for hashing and equiality checks.

Another major change is to use the default `magma_coreir_context`
instead of using a new context for the uniquification compiling logic.
The issue here was the use of wrapped circuits which were already
defined w.r.t. the default magma context.  There were issues related to
compiling multiple times within the same context, but they have been
fixed with https://github.com/rdaly525/coreir/pull/656 (the issue was
that the JSON pass not being re-run because the graph was not marked as
modified).

Updates the coreir backend to use proper syntax of logging.

Adds the ability to set the database backend (for mantle's verilog
target, it makes more sense to use the magma verilog backend, this
shouldn't affect most user code since the mantle verilog target isn't in
heavy use).

Update the transform logic to end the circuit definition at the end.
Before, it was calling EndCircuit right away (but then populating the
definition later) which was causing issues in the uniquification logic
(can't compile when EndCircuit was called because the definition wasn't
ready).