Relax condition for rewriting select path for nested type
Before the logic was only rewriting the select path when an array or
tuple object was reference from an Array or Tuple. This relaxes it
to always rewrite the select path when any object is referenced from
an Array or Tuple.

In the case of https://github.com/phanrahan/magma/issues/415, a Bit was
being selected from a Tuple, but was not rewritten because it failed
the condition:

    isinstance(arrOrTuple, ArrayType) or \
    isinstance(arrOrTuple, TupleType)

However, regardless of the type of the child object that is being
selected from the tuple/array, we should be rewriting the select path
to flatten out the tuple/array