Fix bug in tuple iswhole logic for bulk connections
This fixes the case where the compiler improperly tries to do a bulk
connection.

The desired behavior is to check:
is the value connected to the current tuple "whole", which means
that a tuple of the same type is connected to the current tuple,
in this case we can  connect the two tuples together, rather than
their components.

However, this check was improperly indicating a bulk connection in the
case when the driving value satisfied all the properties for a whole
tuple, except that it doesn't contain all the keys of the parent tuple.

This change strengthens the check to ensure that the driver is only
considered whole when the values refer to all the values corresponding
to the parent tuple.