Refactor Port logic to use Wire representation
Before, magma represented circuits as direct source -> sink connections
with no intermediates.  When a temporary value was wired to another
value, their connections were merged and temporary effectively removed.

In preparation for the logic that will preserve wire names in the
generated code, these changes refactor the internals to preserve
the intermediate connections.  This is done by changing the concept
of a Port into a Wire that maintains a notion of a `driver` and a
list of Wires that it is `driving`.  Wires are qualified by the
bit that they correspond to and use the same logic.

The function of `trace` and `value` were adjust slightly for this, which
required updating the tests but still works as expected for the
internals.  Before, `trace` and `value` operated in certain ways for
inputs and outputs, however the notion of tracing an output is not
used internally anywhere except in the test.  To make the logic more
consistent/precise, `trace` will return the source input/inout driver
of a wire, while `value` will return the directly connected driver.

There's a slight confusion with the name `Wire`, since to me that could
correspond to a bidirectional entity, but here they have a notion
of being an input, output, inout, or neither.  I'm not sure if there's
a better name.  Port worked for this because we never realized the
intermediate values, so we always connected one port to another
(and these were qualified).  Now we have intermediate Wire objects
that have no direction, which end in directed Wires.  Perhaps
the best thing to resolve this would be to introduce a Port object
that is only used by interfaces, and wire objects for temporary values.