Add staged placer for new io syntax
Added a StagedPlacer class to enable the new io syntax. The way this
works is by first using a stack of Placer's instead of Definition's.
Then, we add a __prepare__ method to CircuitKind (and therefore to it's
children classes, e.g. DefineCircuitKind), which pushes a new
StagedPlacer on to the placer stack. This StagedPlacer captures all
instanced circuits, but doesn't place them. This is needed because at
this time the definition doesn't even *exist*. Inside of
CircuitKind.__new__, we consume the instances staged by the staged
placer and actually place them by calling StagedPlacer.finalze(). This
staged placer is then popped off the stack. If we need to enter the
method body of 'defintion()` then we can again push the class's placer
object onto the stack, and operate as usual.