Fix wiring errors in new style syntax
This fixes a problem with the LazyDefnRef approach to ports
in the new syntax. Since 'defn' is None at definition (i.e. wiring
time), error msg's which need the qualified/debug name of a port
would fail because the containing defn doesn't exist. This is
actually a fundamental problem with this approach: the defn doesn't
even exist at wiring time, and may not even be named yet. To fix
this, we set defn to be a dummy object with a single attribute
'name' which is just "". That way these code paths will work but
will return an emtpy name for the containing defn.

Also fixes a bug where __name__ was used instead of name.