Add prototype for NDArray
This minimum working example subclasses the Array/ArrayMeta classes to
minimize the amount of code to write.  One problem with this approach is
that the Array/ArrayMeta logic depend on the `Array[N, T]` interface for
dealing with various magma operations such as flipping types.  The
NDArray type does not satisfy this interface for it's type constructor,
so for now it just assumes invalid parameters are to be passed through
to the Array super class logic.  The downside to this approach is that
in theory, a user could do `NDArray[N, T]` and get back an `Array[N, T]`
rather than error message (for now NDArray parameters should be a tuple
of values convertible to int).

Two options I can immediately see for cleaning up this pattern:
1. Move the generic Array/ArrayMeta logic into a base class, inherit
from Array and NDArray.  I'm not exactly sure how this would work, but
this seems like the natural choice for reusing the core logic code with
different interfaces.  I have a feeling this will be an non-trivial
undertaking just based on the complexity of the Array/ArrayMeta logic.
2. Use the MagmaProtocol pattern for NDArray so the underlying value
uses the Array logic and the NDArray interface just dispatches
conveneniences like tuple indexing to the underlying value.  This seems
like a simpler choice that doesn't involve messing with the
Array/ArrayMeta pipeline