Add preliminary support for BitPattern
This adds support for declaring and comparing bit pattern values that
can include don't cares (?) ala chisel's BitPat:
https://github.com/freechipsproject/chisel3/blob/v3.3.2/src/main/scala/chisel3/util/BitPat.scala

For now, we implement the comparison logic in magma using a mask for the
don't care bits, in the future we should consider whether it matters
that we generate these in the output verilog (for synthesis quality or
readability).  If it does matter, we'll need to add support for
compiling it through CoreIR.  This uses the chisel syntax for bit
pattern strings.