Implement quick fix for inheriting Product types
Product types defined by multiple layers of inheritance, e.g.

  class ParentProd(m.Product): ...
  class ChildProd(ParentProd): ...

are tricky because `ChildProd._unbound_base_ = None`. Therefore
we need to recurse up the inheritance tree to find the closest
unbound base.

NOTE: I am not sure if this is the best way to do this, but it works
for the simple cases of interest.