Fix bug with automatic wiring of temporary clock
This adds a test and fixes bugs related to automatic wiring of temporary
clocks.  The issue is that the automatic clock wiring logic assumed that
clocks were either driven or undriven at the sink of an instance.
However, in some case, the clock could be driven by a temporary value,
which itself isn't driven.  https://github.com/phanrahan/magma/pull/823
addressed one case where this issue arose, but did not cover the cases
presented in the added test.

First, the InsertWrapCasts logic had to be improved to traverse the
drivers of a port to find any instances where wrapping was necessary.
For example, we could have a temporary value of type Bit driving an
instance port of type Bit (no wrapping needed).  However, the driver of
the temporary Bit might be a clock, so a wrap needs to occur there.

The repr/insert_coreir_wire logic has been improved to handle anon bits
that are no driven (before it would try to call value on these to "skip"
them, but since they are undriven value would actually return the value
that it drives).

This changes the internal `driven` API to return True if a value is
driven (something is wired up to it), rather than True if trace() is
None.  The idea here is that this new version of `driven` is useful for
the above changes, while when we want to check whether a value can be
traced to a source driver, we can simply do trace() is None (whereas
there was no API to check if something is driven without tracing all the
way to a source).