Add built-in Valid type
This provides a standardize type for the valid pattern.  It uses the
same getitem interface as the other parametrized types (e.g. Valid[T]).
It creates a product with two fields:
* `valid: Bit`
* `data: T`

For now it doesn't provide any other methods but we can use the Valid
class as a place to add any useful functions.  We can also start using
introspect on Valid types to define standard/reuseable code based on
Valid interfaces.