Restrict inline_combinational to use wiring
Avoids the problem discussed in https://github.com/phanrahan/magma/issues/812

Since we cannot guarantee that the assignment values can be used in the
outer scope, we restrict that the user use wiring.  If the user wants a
value to "escape" to the outer scope, they can create a temporary value
and wire up to it.

Also fixes a wiring bug in the inline comb logic (it should be wire
driver to drivee, this matters for using temporary values)

Fixes #812