Add ReadyValid type
* Based on
  https://github.com/freechipsproject/chisel3/blob/v3.3.2/src/main/scala/chisel3/util/Decoupled.scala
* We use decoupled.py since we plan to add the other related types
* We add a notion of Produce/Consume type modifiers for clarity

The ReadyValid methods use a "last connect" style semantics borrowed
from chisel.  This allows the user to set "default" values for wires and
override their values based on conditions.  The ordering of the method
invocations set the priority for the conditions.  This emulates the
chisel "when" statement semantics using a method interface rather than
implicitly capturing this behavior when methods are invoked inside
conditional statements.

In a subsequent PR I plan to add support for these methods inside
combinational which will dispatch to these underlying methods, but for
now we can expose them for normal use (as shown in the test).  The main
value in adding combinational support is we can combine multiple enq/deq
calls with the same condition, e.g.
```python
if cond:
    a = x.deq()
    y.enq(b)
    z.enq(w.deq)
```