Restrict Clock wiring
Fixes https://github.com/phanrahan/magma/issues/651

Enforces nominal typing for clock types. Clocks (and
resets/enables/etc...) can only be wired to values that are instances of
the same type (i.e. not instances of Digital).

Includes updates to portions of the code base that were relying on the
ability to wire Digital/Bit to clock types.  Mainly the change requires
using an explicit `m.clock` or `convertbit(value, T)` (for a generic
conversion of value to T) to connect a bit to a clock type.