Make register init/coreir_init consistent
Before, if we had a synchronous reset or no reset, we skipped setting
the coreir_init variable (set it to 0 instead).  I think the reasoning
may have been with a synchronous reset, the init value is also provided
to the mux logic.  This changes the behavior so that the coreir init
matches the actual init in the sync reset or no reset cases.  I can't
think of a reason why we shouldn't do this, whereas the current behavior
provides unexpected results (e.g. if I don't specify a reset, the
register init value is just completely ignored, rather than being used
to initialize the register contents)