Fix ite type checking bug (#571)
Before, ite was allowing SInt and UInt to be used together, which should
not be allowed.  This fixes it by first updating the ite logic to use
the mux operator (which has improved type checking logic) and updating
the wireable logic to prevent UInt and SInt from being wireable (they
match structurally but we should error when trying to wire them
together).  I think we should eventually update the wireable logic to
allow the types to define their wireability themselves (rather than
housing the logic internal to magma since this won't extend to user
defined types).  However, since UInt/SInt are built-in to magma, this
should be fine for now.