Fix CoreIR name error for inline verilog temporary
Before, inserting a temporary wire for the inline verilog logic could
trigger a CoreIR name error by using a directed type for the wire name
(e.g. `Out(Clock)`) since the Wire genertor logic generates a name
`"Wire{T}"`.  It doesn't really make sense to generate a Wire of type
`Out(T)`, rather it should be some undirected type `T`.  This updates
the Wire generator raise an error if you try to construct it with a
directed type, and updates the inline verilog code to use the undirected
type for the intermediate wire.