[seq2] add magma reference if not available in env
Before, we searched for a reference to magma in the sequential2
environment (we need this to add references to types when generating the
reset code).  If we didn't find one, we raised an error.  This changes
the logic to insert a reference to magma with a free name when we can't
find an existing one.