Fix bind bug with ndarrays on nested instances
Before, the bind logic was not properly handling ndarrays found in
nested instances.  This because the type of the values was PortView
rather than Array, so the various checks were failing to detect the
underlying type.  This changes the logic to unpack PortView values into
their underlying values for these checks.