Add utility fns to convert from magma to hwtypes
Intended usecase is for fault where we inspect the magma circuit type
and automatically martial values into the corresponding hwtypes type for
simulation.

Note for now we don't support nested arrays since we don't have an
analagous hwtypes type.  In theory we could use numpy arrays and set the
datatype to be the leaf hwtypes type.