[Fix #898] Raise exception when circuit has errors
This adds logic to the staged logger flow to store whether any error
messages were recorded during the circuit definition.

If an error was encountered, the backend will raise an exception rather
than proceeding.

This should avoid difficult to parse errors in the backend when
operating on a malformed circuit.  This should also hopefully surface
the error message to the user without less visual noise (e.g. a
traceback from an exception in the backend, leading them to believe it
could be an internal compiler error or something).

The hope is that if a user doesn't want something to be an error, then
it should be explicitly marked by the user.  E.g. if they're leaving a
value undriven on purpose (so a subsequent pass can drive it), then they
should mark it with `undriven()` and the downstream pass should be
responsible for handling the undriven value.  This is the less common
case, so I think it's better for these users to have to explicitly do
things to avoid errors (so the common case when something is an error is
caught by default).