Avoid duplicate runs of InstanceGraphPass
Previously, in the CoreIR backend, DefinitionTransformer was running
the instance graph pass every time. In reality, we only need to run it
for the root (top) circuit, and children nodes can skip it.

Rather than re-architecting the entire flow to build it explicitly
once, we add an option to skip the instance graph pass, which is by
default False, and set to True by the root node.