Fix error with VCC
Before this change, the added test would produce this error:
```
>       raise NotImplementedError(ref, type(ref))
E       NotImplementedError: (VCC, <class 'magma.ref.NamedRef'>)
```

The issue is that ref is not an AnonRef (where const values are
handled), rather it's a NamedRef.  This changes it so that if a value is
const (regardless of the type of the ref), it should work.