Add inline instance API
This change just adds the APIs and nec. data structures for keeping
track of inlining. In particular, we add an extra field to the symbol
table for such entries. We also add a sentinel value for keeping track
of instances which have been inlined, as well as a serialized (json)
representation.

NOTE: This change *does not* implement the logic needed to track
inlined instances in the master (chained) symbol table.