Add support for verilog_name metadata
CoreIR has a metadata field "verilog_name" that allows a module to
explicitly set a desired verilog name.  It's used for the ice40
primitives, which are defined in the ice40 namespace.  Normally, you'd
get something like ice40_<MODULE_NAME> given the namespace to verilog
name logic, but the verilog_name metadata overrides this logic and fixes
a specific string.  This metadata also overrides the verilog_prefix
logic, which gives the user the ability to mark certain modules to be
ignored by the verilog_prefix.  This is useful when you want most of
your design to use the verilog_prefix, except a select few modules that
you don't control the name of (e.g.  standard cells).