Fix error with reset cast in python sim flatten
Before, we'd get this error:
```
ERROR:magma:Cannot wire Foo_flattened.RESET (Out(Reset)) to coreir_commonlib_mux2x4_inst0.in.sel[0] (In(Bit))
```

This is because inside the register code, we cast a register to bit (as
an input to a mux).

This fix changes the flattening code to cast a new source bit before
wiring it to a sink bit since at this point we can assume the user code
was valid (i.e. they performed a cast).