Support new instance name type
The API has changed from the type of an instance name being a sum (
python "Union", c++ "variant") to being a tuple of (Sentinel, string).
This now encodes all cases, where either the name is valid (sentinel ==
empty), or the name is a key used for inline lookups (sentinel ==
inlined).