Optimize InstView logic
This was one source of the memory issues.  The sub instances can share
the same reference to their parent rather than creating a unique one for
each child.  We can improve this further, but this is a simple stop gap.
To improve it, we should make this hierarchical reference logic lazy
(add getattr logic that fetches these sub instances on demand) rather
than building up the references ahead of time (since the common case is
to not use these references).  The problematic circuit had 10,000
instances, causing this to become quite an issue.