Raise exception when trying to compile empty defn
This came up with an old-style class definition that had a typo in the
`definition` method (spelled it `definiiton`).  This resulted in a
segfault in coreir since it was trying to compile an empty top.  We
should improve the coreir error handling, but I think at this level it's
fine to raise an Exception when the user tries to do this, because
there's not really any reasonable use case for trying to compile an
empty circuit without a definition (we could just generate a verilog
declaration? but seems like an odd use case for magma that we don't need
to support)