Share inline wires for undriven clocks
The current pattern for using the default clocks in inline verilog is to
refer to an undriven anonymous clock (e.g. `clk = m.Clock()`).  Then,
the compiler will use the default clock driver logic to wire this up.
This is useful for the pattern where you have a helper function for
generating properties, with an optional clock argument (so the user can
pass a specific clock, or by default it uses the default clock).  Since
this helper function doesn't know the context in which it's being
invoked (i.e. the default clock of the definition it's being used in),
it simply uses an undriven anonymous clock that will be automatically
wired up later.

The current inline_verilog logic will generate multiple wires for each
undriven clock because they're different anonymous values, which is
annoying (since many properties use the default clock, we get many wires
for the default clock, but they are all the same clock).

This changes the logic to handle undriven clocks and share a wire for
each undriven clock type, with the intention that they will all use the
same default driver wired up later.  Note that we don't handle general
(non clock types) since this logic happens during the process inline
verilog pass during the compilation phase.  At this point, we assume
that all values have been wired up, so if there's no driver for a
non-clock value, then it's an error (won't be handled by the automatic)
clock wiring logic, so we only need to consider the case where clock
types are undriven.