Add support for name argument in mux operator
Since kwargs isn't being used (only T and height are arguments to the
Mux generator), we swap the logic to passthrough kwargs to the instance
to pass any instance parameters.  In the future we could special case
the name kwarg if we wanted to pass through kwargs to the generator, or
vice versa (special case generator kwargs).