Add default clock if none is present
Fixes #970

There's a common case where the user forgets to specify a clock port
with m.ClockIO.  The requested feature is that in this case, we
automatically lift the clock port to the external interface.

I investigate a few options for doing this.  The first idea was to do it
in the backend using a pass.  The problem with this is at this point the
IO is already bound, and cannot be modified.

The cleanest solution I was able to arrive at was during circuit
construction time.  Before we bind the IO (and it can no longer be
modified), we check whether there's any clock types.  If not, we add the
default clock port, then the standard automatic clock wiring logic will
take things from there.  If the added clock port is unused, (i.e.
there's no instances that require it), we remove the port in the
backend.