Fix issue in flip then qualify of ready valid
Currently, calling qualify on a ready valid type (e.g. `In(T)`) will
behave correctly in that it will return a type that has all the
subfields (ready, valid, data) as inputs.  However, calling flip on the
qualified type will actually return the Flip of the original type.  This
is because producer/consumer ready valid types define flip to return the
corresponding dual.  The problem here is that when we call qualify using
the default Type.qualify, it returns a Producer/Consumer type with the
correct field types, but doesn't change the flip behavior.

The proposed change is that when you call qualify on a readyvalid type,
you get an AnonProduct back with the qualified fields that behaves like
a normal tuple.  This seems desirable since once you change a readyvalid
type to be qualified as all In or Out, it no longer really behaves like
a standard ReadyValid type and therefor doesn't obey the interfaces
(e.g. enq/deq, however there might be some use case for calling fired on
it but that could be defined as as helper function).

The use case for this pattern is inside the bind/monitor flow which
takes a generic circuit and converts all the ports to inputs (so the
monitor can read the IO).  Calling In on a ReadyValid type works fine,
but calling Flip on it (inside the Circuit interface flow) produces the
wrong behaviors (returns a Producer or Consumer, instead of something
with all inputs/outputs)