Resolve regression with ReadyValid.undirected_t
This is used in the product/tuple base class logic (see
https://github.com/leonardt/magma_riscv_mini/runs/3216036083?check_suite_focus=true
for a failure example).  I think it is reasonable to be able to get the
"undirected" version of a ReadyValid and the recent change to the
monitor logic removed this capability.  These changes add back the
capability while preserving the prevention of using the In/Out qualifier.