Fix infinite loop in repr logic
Previously the repr logic for emitting wires stopped when encountering
an output value.  This assumes that all values are eventually traced
back to source a driver.  In a case where there is no source driving an
undirected temporary, the value API will actually return the drivee,
causing an infinite loop where the the logic keeps emitting wires.

In this case, what happens is the wire tracing encounters an anonymous
tuple value created by the combinational logic.  In this case, the
anonymous tuple is not driven, since it's comprised of output members,
but the directionality is not an output since it's a "temporary" value.

This fixes the issue by checking if the driver is itslf driven, rather
than if it is an output, and stopping the "trace" for emitting wires
when encountering an undriven value.

Another option would be to update the tuple logic to mark this as an
output rather than a temporary type.  However, I think this fix covers
another possible case in which the user leaves a temporary value
undriven, in which case it would infinite loop in the same way.  Now
this might not be a problem if it's a valid circuit to compile (e.g.
they used the undriven pass), but this would avoid an infinite loop if
the called repr before the compiler passes are run which might be useful
if printing out the circuit during incremental development (e.g. want to
dump the repr of an incomplete circuit definition to get an overview of
the current structure)

Fixes #995