Fix bug in Array2 wiring logic for anon values
Before, if an array was driven by an anon value, we did a bulk wire,
however this doesn't maintain consistency properly for the children of
the anon value.  I.e., the anon value will populate the driving field
for the bulk array, but the children will not have their driving field
updated.  To remedy this, if we are being driven by something with
elaborated children, we perform a recursive wire to maintain consistency
of those children's wiring interface.