Use reduce logic for Array equality
Using equality of lists (since .ts returns a list of children) will recursively
call equals on the members, but then do a reduction to determine if they're all
equal.  This is problematic in magma since this will call `__bool__` on the
output of the equal operator on the children.  Instead we want to explicitly
perform an and reduction.

This is an example of the error we get with the current logic:
```
tests/test_type/test_array.py:222: in Foo
    io.O @= io.I0 == io.I1
magma/operator_utils.py:12: in _wrapped
    return fn(self, *args, **kwargs)
magma/array.py:410: in __eq__
    return self.ts == rhs.ts
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = .out

    def __bool__(self) -> bool:
        if not self.const():
>           raise ValueError(
                "Converting non-constant magma bit to bool not supported")
E           ValueError: Converting non-constant magma bit to bool not supported

magma/digital.py:234: ValueError
```