Fix bug in enumerate children slice index logic
Before, we assumed the value of the slice was itself a slice object, but
after the changes in https://github.com/phanrahan/magma/pull/1045 this
is no longer true since an Array can be constructed with an
existing value as the slice.

The included shows one such case, where two slices are wired up to
InstRef values, and calling trace() returns an object where the slice
value is an instref.

This updates the logic to store the original slice_ key with the slice_
value in the data structure used for the slice lookup.