Fix bug in overlapping slice resolution
Follow up to https://github.com/phanrahan/magma/pull/1045 and
https://github.com/phanrahan/magma/pull/1048

Because the collect_children logic now creates Arrays where the slice children
could be from other values, we need the slice resolution logic to decide
whether the slice value is from the current array (e.g. resolving x[0:2] for x)
or from another value (e.g. resolving y for [0:2] of an anon value).

This hotfixes the build issue on master for FB, but will need to add a unit
test to cover this case.