Fix bug in wire children logic missing unwire (#1056)
* Fix bug in wire children logic missing unwire

Before, we were relying on the "overwriting" log message when wiring a
second driver to a child.  However, this is problematic because we don't
actually unwire the child being overwritten.  As as result, in various
code paths (e.g. error reporting) we can trigger one of the overwritten
drivers to expand itself, which in turn calls the
_resolve_driven_bulk_wires logic and corrupts the representation.  By
unwiring the overwritten driver, we avoid any code acciddentally
triggering the resolve logic.

* Add driven check

* Add broken temp test

* Remove redundant check

* Generalize fix to wiring logic

* Remove old code

* Remove trailing comma