Avoid unnecessary array expansion in repr
Noticed this in debugging, before this change we were getting:
```
        wire(Foo.I.c[1].a[3][0], Foo.O.c[0].a[0][0])
        wire(Foo.I.c[1].a[3][1], Foo.O.c[0].a[0][1])
        wire(Foo.I.c[1].a[3][2], Foo.O.c[0].a[0][2])
        wire(Foo.I.c[1].a[3][3], Foo.O.c[0].a[0][3])
        wire(Foo.I.c[1].a[2][0], Foo.O.c[0].a[1][0])
        wire(Foo.I.c[1].a[2][1], Foo.O.c[0].a[1][1])
        wire(Foo.I.c[1].a[2][2], Foo.O.c[0].a[1][2])
        ...
```

The issue was in two places: first, we were recursing before traversing
the anon temporary values (causing expansion), and second in the logging
we were calling repr on an unelaborated anon array which was triggering
the elaboration logic by default.  Instead, we only emit the `array([t0,
t1, ...])` style if we have elaborated children, otherwise we fall back
to the normal value repr (emits automatic anon name).