[MLIR] Add undriven support (#1077)
* [MLIR] Make sv.WireOp 'name' attribute optional

* [MLIR] Add tests for coreir.undriven primitive

* [MLIR] Add support for coreir.undriven primitive

* [MLIR] Attach symbol to undriven wires

We attach an (MLIR) symbol (sym) to the SV wires emitted for
coreir.undriven instances. This ensures that the wire is not
transformed or optimized away (usually by canonicalization which
converts dangling wires to X's).

Note that we *don't* specify the name since we typically don't care
about the dangling wire name and want to let the backend use
temporaries.