Fix rebase bug in aggregate wireables, add tests
I chose the wrong changes for these functions when rebasing the tuple2
code.  With the aggregate_wireable_method decorator, they only need to
implement the logic for when children are elaborated (the decorator
calls the aggregate method if the has_elaborated_children check fails).
This didn't trigger a CI bug because when unwire fails, it just raises a
warning (e.g. driving an already driven value) that occurs silently in
the passing tests.  Adds a test to cover this case.