Use AggregateWireable API to avoid elaboration
Some cases in the repr and insert_wrap_casts logic were unnecessarily
expanding arrays.  This updates the logic to avoid elaboration when
possible, also updating it to use the AggregateWireable API.

Some of the smart_bits tests relied on this elaboration for checking the
repr of two circutis, so some code has been added there to force
elaboration so the repr works.