[Primitives] Adds default=None for list_lookup
Rather than use an explicit default value as the standard pattern, it
makes more sense to use the final element of the list since we're
buildling a mux tree anyways.  The existing explicit "default" value
pattern might still be useful to emulate a certain "else" pattern that's
different than the last element.

In adding this, I noticed the mux tree was being built forwards, which
made the generated code a bit convoluted for this default case, so I
reversed the tree instead to make it look more natural in the generated
code.